(misc) move provtarget package under providers
