+++ go.mod
@@ -27,8 +27,8 @@ require (
 	github.com/looplab/fsm v0.2.0
 	github.com/miekg/pkcs11 v1.0.3
 	github.com/mitchellh/mapstructure v1.4.1
-	github.com/nats-io/jsm.go v0.0.26-0.20210805101059-aa3de0ef564b
-	github.com/nats-io/nats-server/v2 v2.3.4
+	github.com/nats-io/jsm.go v0.0.26-0.20210809070242-d17e3f1dc702
+	github.com/nats-io/nats-server/v2 v2.3.5-0.20210808150006-52cfce24c426
 	github.com/nats-io/nats-streaming-server v0.22.1 // indirect
 	github.com/nats-io/nats.go v1.11.1-0.20210803204434-91bdffe39f41
 	github.com/nats-io/stan.go v0.10.0
