+++ providers/provtarget/provtarget.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 
 	"github.com/choria-io/go-choria/config"
-	"github.com/choria-io/go-choria/provtarget/builddefaults"
+	"github.com/choria-io/go-choria/providers/provtarget/builddefaults"
 	"github.com/sirupsen/logrus"
 
 	"github.com/choria-io/go-choria/srvcache"
