+++ providers/provtarget/provtarget_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/choria-io/go-choria/build"
-	"github.com/choria-io/go-choria/provtarget/builddefaults"
+	"github.com/choria-io/go-choria/providers/provtarget/builddefaults"
 	"github.com/golang/mock/gomock"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
